requirements.txt
alembic==0.8.10
appdirs==1.4.1
Mako==1.0.6
MarkupSafe==0.23
packaging==16.8
PyMySQL==0.7.10
pyparsing==2.1.10
python-editor==1.0.3
six==1.10.0
SQLAlchemy==1.1.5
websocket-client==0.40.0
pymongo==3.7.2
numpy==1.16.2
wsaccel==0.6.2
pyyaml==3.13